Move basic actions to separate directory
